requirements.txt
streamlit
whoosh
python-docx
python-doc
pdfminer-six
requests
sentence-transformers
faiss-cpu
pymupdf
pytesseract
pdf2image
pillow
python-dotenv
openai==0.28
